requirements.txt
fastapi
uvicorn
requests
pytest
streamlit
python-dotenv